refactor: a lógica de UI deixa de conhecer o motor, e a ABI antiga vira feature
rts-egui, rts-dom, rts-render e rts-input dependiam de `rts-engine` e, por ele,
de `rts-abi` — a interface que `rts_cranelift::abi` substituiu. Um crate do
motor novo que os alcançasse alcançaria, pelo grafo de build, exatamente o que
está sendo removido; então nenhum deles podia ser consumido pelo motor novo.

Agora a lógica é Rust comum (`&str`, `f64`, `u64`) e a casca `extern "C"` do
motor antigo vive num módulo próprio atrás da feature `old-engine`, ligada por
default para que o motor antigo siga compilando sem mudança. `rts-runtime`, que
É o motor antigo, passa a pedi-la explicitamente.

Feature e não `cfg` solto porque a pergunta é de DISPONIBILIDADE — "este build
tem a ABI antiga?" — que é a única razão que `docs/engine/architecture.md`
aceita para uma fronteira de crate.

Duas coisas que não são casca, e por isso ficaram atrás da feature em vez de
migrarem: `rts:gpu` (`compute.rs`), cujos buffers SÃO `Entry::Buffer` do
`HandleTable` antigo, e `drawWater`, seu único consumidor. Portá-las é escolher
onde aqueles bytes vivem no motor novo, e isso é decisão da fase seguinte —
dita aqui em vez de descoberta por um link quebrado.

`crate::pixels` é o mesmo problema resolvido para a `<img>`: o render pergunta
de onde vêm os bytes em vez de saber. O motor antigo instala a resposta ao
registrar o namespace; o novo instalará a sua.

Nenhum símbolo `__RTS_FN_NS_EGUI_*` mudou de nome ou assinatura — os 45 que o
`symbol_table.rs` referencia continuam existindo, agora como cascas.
