+++ crates/rts-egui/src/abi/register.rs
@@ -1,0 +1,357 @@
+//! A tabela do namespace `egui` do motor ANTIGO: nome, assinatura, símbolo.
+//!
+//! Separada de `super` porque as duas coisas mudam por motivos diferentes — um
+//! wrapper muda quando a função muda, esta tabela muda quando a SUPERFÍCIE
+//! muda — e porque juntas passavam do teto de arquivo.
+
+use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
+
+// `Sig::new` espera variantes de `AbiType` como valores. Aliases locais para
+// manter as tabelas de membros legíveis (mesmo shape do `io::func`).
+use AbiType::{F64, I64, U64};
+
+use super::*;
+
+/// Helper de declaração de membro do namespace (mesmo shape do `io::func`).
+fn func(name: &str, symbol: &str, sig: Sig, ts: &str, doc: &str, fp: *const u8) -> Member {
+    Member {
+        name: name.to_string(),
+        kind: MemberKind::Function,
+        sig,
+        symbol: symbol.to_string(),
+        fn_ptr: FnPtr(fp),
+        flags: MemberFlags::NONE,
+        aliases: Vec::new(),
+        variadic: false,
+        ts_signature: ts.to_string(),
+        doc: doc.to_string(),
+        ret_class: None,
+        pure: false,
+        emit: None,
+    }
+}
+
+/// Registra o namespace `egui` no motor antigo. Resolve via Registry (doutrina
+/// PRIMORDIAL-vs-Registry): o engine NUNCA nomeia `ui`.
+pub fn register(e: &mut Engine) {
+    // Registra o egui como o backend de render ATIVO do `rts-render`: a partir
+    // daqui, o namespace `render.*` (que o DOM/layout chama) despacha p/ o egui.
+    crate::render_backend::register_backend();
+    // De onde o render lê os pixels de uma `<img>`: do `HandleTable` deste
+    // motor. O crate de render não nomeia mais o HandleTable (ele serve aos dois
+    // motores), então quem sabe responde — ver `crate::pixels`.
+    crate::pixels::set_source(|handle, offset, len| {
+        rts_engine::heap::handles::with_entry(handle, |entry| match entry {
+            Some(rts_engine::heap::handles::Entry::Buffer(bytes)) => {
+                let start = offset as usize;
+                (bytes.len() >= start + len).then(|| bytes[start..start + len].to_vec())
+            }
+            _ => None,
+        })
+    });
+    e.ns("egui")
+        .doc("Immediate-mode GUI primitives (egui). TS drives the render loop.")
+        // ── Ciclo de vida da janela / loop (Modelo A — TS dirige) ──────────────
+        .member(func(
+            "openWindow",
+            "__RTS_FN_NS_EGUI_OPEN_WINDOW",
+            Sig::new(vec![AbiType::StrPtr, I64, I64, I64], U64),
+            "openWindow(title: string, w: number, h: number, config: number): number",
+            "Opens a window + render backend, returns an opaque UiCtx handle. `config` is a GPU bitfield (bit0 high-power, bit1 perf-memory, bit2 high-limits); 0 = RAM-optimized defaults.",
+            __RTS_FN_NS_EGUI_OPEN_WINDOW as *const u8,
+        ))
+        .member(func(
+            "pump",
+            "__RTS_FN_NS_EGUI_PUMP",
+            Sig::new(vec![U64], I64),
+            "pump(h: number): number",
+            "Pumps pending OS events (non-blocking). Returns 0=continue, !=0=exit.",
+            __RTS_FN_NS_EGUI_PUMP as *const u8,
+        ))
+        .member(func(
+            "isOpen",
+            "__RTS_FN_NS_EGUI_IS_OPEN",
+            Sig::new(vec![U64], I64),
+            "isOpen(h: number): boolean",
+            "Returns true while the window has not been closed.",
+            __RTS_FN_NS_EGUI_IS_OPEN as *const u8,
+        ))
+        .member(func(
+            "close",
+            "__RTS_FN_NS_EGUI_CLOSE",
+            Sig::new(vec![U64], AbiType::Void),
+            "close(h: number): void",
+            "Destroys the window and frees the UiCtx.",
+            __RTS_FN_NS_EGUI_CLOSE as *const u8,
+        ))
+        .member(func(
+            "moveWindow",
+            "__RTS_FN_NS_EGUI_MOVE_WINDOW",
+            Sig::new(vec![U64, I64, I64], AbiType::Void),
+            "moveWindow(h: number, x: number, y: number): void",
+            "Moves the window to absolute desktop position (x,y) in physical pixels — pick a monitor (e.g. x >= primary width = secondary screen).",
+            __RTS_FN_NS_EGUI_MOVE_WINDOW as *const u8,
+        ))
+        .member(func(
+            "setVsync",
+            "__RTS_FN_NS_EGUI_SET_VSYNC",
+            Sig::new(vec![U64, I64], AbiType::Void),
+            "setVsync(h: number, on: number): void",
+            "Runtime per-window vsync toggle. on!=0 = Fifo (the mandatory default); on==0 = explicit opt-out via AutoNoVsync (uncapped fps — the caller owns throttling/CPU burn). wgpu backend only.",
+            __RTS_FN_NS_EGUI_SET_VSYNC as *const u8,
+        ))
+        .member(func(
+            "mouseLock",
+            "__RTS_FN_NS_EGUI_MOUSE_LOCK",
+            Sig::new(vec![U64, I64], AbiType::Void),
+            "mouseLock(h: number, on: number): void",
+            "FPS pointer-lock: on!=0 confines+hides the cursor and switches input.mouseDeltaX/Y to RAW device deltas (edge-immune); on==0 releases. Windows uses Confined+raw (winit has no Locked there).",
+            __RTS_FN_NS_EGUI_MOUSE_LOCK as *const u8,
+        ))
+        .member(func(
+            "setNextWindowPos",
+            "__RTS_FN_NS_EGUI_SET_NEXT_POS",
+            Sig::new(vec![I64, I64], AbiType::Void),
+            "setNextWindowPos(x: number, y: number): void",
+            "Sets the INITIAL position of the NEXT openWindow (physical px) — the window is born there. Call BEFORE openWindow; more reliable than moveWindow.",
+            __RTS_FN_NS_EGUI_SET_NEXT_POS as *const u8,
+        ))
+        // ── Teste visual headless ──────────────────────────────────────────────
+        .member(func(
+            "snapshot",
+            "__RTS_FN_NS_EGUI_SNAPSHOT",
+            Sig::new(vec![U64, AbiType::StrPtr], AbiType::Void),
+            "snapshot(h: number, path: string): void",
+            "Schedules a PPM snapshot of the next endFrame (glow backend only) — for headless visual assertions that the frame is not blank.",
+            __RTS_FN_NS_EGUI_SNAPSHOT as *const u8,
+        ))
+        // ── Frame ──────────────────────────────────────────────────────────────
+        .member(func(
+            "beginFrame",
+            "__RTS_FN_NS_EGUI_BEGIN_FRAME",
+            Sig::new(vec![U64], AbiType::Void),
+            "beginFrame(h: number): void",
+            "Starts an egui pass: takes input and creates the root Ui.",
+            __RTS_FN_NS_EGUI_BEGIN_FRAME as *const u8,
+        ))
+        .member(func(
+            "endFrame",
+            "__RTS_FN_NS_EGUI_END_FRAME",
+            Sig::new(vec![U64], AbiType::Void),
+            "endFrame(h: number): void",
+            "Ends the egui pass, tessellates and presents the frame.",
+            __RTS_FN_NS_EGUI_END_FRAME as *const u8,
+        ))
+        // ── Widgets-folha (P1) ───────────────────────────────────────────────────
+        .member(func(
+            "label",
+            "__RTS_FN_NS_EGUI_LABEL",
+            Sig::new(vec![U64, AbiType::StrPtr], AbiType::Void),
+            "label(h: number, text: string): void",
+            "Emits a text label in the active frame.",
+            __RTS_FN_NS_EGUI_LABEL as *const u8,
+        ))
+        .member(func(
+            "button",
+            "__RTS_FN_NS_EGUI_BUTTON",
+            Sig::new(vec![U64, AbiType::StrPtr], I64),
+            "button(h: number, label: string): boolean",
+            "Emits a button; returns true if it was clicked this frame.",
+            __RTS_FN_NS_EGUI_BUTTON as *const u8,
+        ))
+        .member(func(
+            "slider",
+            "__RTS_FN_NS_EGUI_SLIDER",
+            Sig::new(vec![U64, F64, F64, F64], F64),
+            "slider(h: number, value: number, min: number, max: number): number",
+            "Emits a slider; returns the (possibly updated) value.",
+            __RTS_FN_NS_EGUI_SLIDER as *const u8,
+        ))
+        // ── Layout horizontal (widgets lado a lado) ────────────────────────────
+        .member(func(
+            "horizontalBegin",
+            "__RTS_FN_NS_EGUI_HORIZONTAL_BEGIN",
+            Sig::new(vec![U64], AbiType::Void),
+            "horizontalBegin(h: number): void",
+            "Opens a horizontal scope: widgets until horizontalEnd sit side by side.",
+            __RTS_FN_NS_EGUI_HORIZONTAL_BEGIN as *const u8,
+        ))
+        .member(func(
+            "horizontalEnd",
+            "__RTS_FN_NS_EGUI_HORIZONTAL_END",
+            Sig::new(vec![U64], AbiType::Void),
+            "horizontalEnd(h: number): void",
+            "Closes the most recent horizontal scope; widgets stack vertically again.",
+            __RTS_FN_NS_EGUI_HORIZONTAL_END as *const u8,
+        ))
+        // ── Render do DOM ──────────────────────────────────────────────────────
+        // `render(win, domHandle)`: RENDER GENÉRICO — pinta um DOM do `rts-dom`
+        // (headless, por handle) na janela. O egui só LÊ o DOM e desenha; não o
+        // detém. É o caminho desacoplado (o DOM vive no rts-dom; o egui é um
+        // consumidor trocável). `html(win, string)` é o caminho LEGACY (egui
+        // parseia e detém o próprio DOM) — mantido por compat.
+        .member(func(
+            "render",
+            "__RTS_FN_NS_EGUI_RENDER",
+            Sig::new(vec![U64, U64], AbiType::Void),
+            "render(win: number, domHandle: number): void",
+            "Paints a rts-dom DOM (by handle) into the window — egui only reads the DOM and draws; the DOM lives in rts-dom (headless). Inverts the dependency: any backend consumes the same domHandle.",
+            __RTS_FN_NS_EGUI_RENDER as *const u8,
+        ))
+        .member(func(
+            "html",
+            "__RTS_FN_NS_EGUI_HTML",
+            Sig::new(vec![U64, AbiType::StrPtr], AbiType::Void),
+            "html(h: number, html: string): void",
+            "LEGACY: parses an HTML string and emits widgets (egui owns the DOM). Prefer `render(win, domHandle)` with a rts-dom DOM (headless, decoupled).",
+            __RTS_FN_NS_EGUI_HTML as *const u8,
+        ))
+        // ── DOM/estilo: REMOVIDOS do namespace egui (extração headless) ─────────
+        // defineBlock/defineStyle/defineInline + querySelector/setText/setAttr/
+        // createElement/appendChild/removeNode/domDump migraram 100% para o
+        // namespace `dom` (rts-dom). O egui é RENDER GENÉRICO: só LÊ o DOM e pinta
+        // (via `egui.html` / `egui.render`), nunca detém estado de página. Isso
+        // INVERTE a dependência — o DOM não conhece o render; qualquer backend
+        // (egui hoje; web/headless-png amanhã) consome o mesmo DOM. Ver
+        // docs/ui/html-engine + project_dom_headless_vision.
+        // ── CANVAS BURRO: o TS calcula o layout e emite primitivos; o egui pinta ──
+        // Coords/tamanhos em pontos (number); cores 0xRRGGBBAA (number).
+        .member(func(
+            "drawRect",
+            "__RTS_FN_NS_EGUI_DRAW_RECT",
+            Sig::new(vec![U64, F64, F64, F64, F64, I64, F64, I64, F64], AbiType::Void),
+            "drawRect(h: number, x: number, y: number, w: number, h_: number, fill: number, strokeW: number, stroke: number, radius: number): void",
+            "Dumb-canvas filled rect + optional border at absolute coords. Colors 0xRRGGBBAA. The TS layout engine positions it; egui only paints.",
+            __RTS_FN_NS_EGUI_DRAW_RECT as *const u8,
+        ))
+        .member(func(
+            "drawText",
+            "__RTS_FN_NS_EGUI_DRAW_TEXT",
+            Sig::new(vec![U64, F64, F64, AbiType::StrPtr, I64, F64, I64], AbiType::Void),
+            "drawText(h: number, x: number, y: number, text: string, color: number, size: number, flags: number): void",
+            "Dumb-canvas text at absolute (x,y) top-left. flags bitmask 1=bold 2=italic 4=mono. Color 0xRRGGBBAA.",
+            __RTS_FN_NS_EGUI_DRAW_TEXT as *const u8,
+        ))
+        .member(func(
+            "drawLine",
+            "__RTS_FN_NS_EGUI_DRAW_LINE",
+            Sig::new(vec![U64, F64, F64, F64, F64, F64, I64], AbiType::Void),
+            "drawLine(h: number, x1: number, y1: number, x2: number, y2: number, w: number, color: number): void",
+            "Dumb-canvas line at absolute coords. Color 0xRRGGBBAA.",
+            __RTS_FN_NS_EGUI_DRAW_LINE as *const u8,
+        ))
+        .member(func(
+            "measureText",
+            "__RTS_FN_NS_EGUI_MEASURE_TEXT",
+            Sig::new(vec![U64, AbiType::StrPtr, F64, I64], F64),
+            "measureText(h: number, text: string, size: number, bold: number): number",
+            "Width (points) of a text in the real font — the ONLY layout-aware op left in egui (TS can't measure text without the font). Synchronous.",
+            __RTS_FN_NS_EGUI_MEASURE_TEXT as *const u8,
+        ))
+        // ── Pipeline 3D wgpu (scene pass) — só-wgpu; glow no-op ─────────────────
+        .member(func(
+            "meshUpload",
+            "__RTS_FN_NS_EGUI_MESH_UPLOAD",
+            Sig::new(vec![U64, U64, I64, U64, I64], U64),
+            "meshUpload(h: number, vertsPtr: number, vertexCount: number, indicesPtr: number, indexCount: number): number",
+            "Uploads a mesh to GPU (verts = interleaved pos+normal, 6 floats/vertex at vertsPtr; indices = u32 at indicesPtr). Returns a mesh id. GPU 3D scene pass, drawn before the egui UI.",
+            __RTS_FN_NS_EGUI_MESH_UPLOAD as *const u8,
+        ))
+        .member(func(
+            "meshFree",
+            "__RTS_FN_NS_EGUI_MESH_FREE",
+            Sig::new(vec![U64, U64], AbiType::Void),
+            "meshFree(h: number, meshId: number): void",
+            "Frees a mesh uploaded with meshUpload.",
+            __RTS_FN_NS_EGUI_MESH_FREE as *const u8,
+        ))
+        .member(func(
+            "setCamera",
+            "__RTS_FN_NS_EGUI_SET_CAMERA",
+            Sig::new(vec![U64, F64, F64, F64, F64, F64, F64, F64], AbiType::Void),
+            "setCamera(h: number, camX: number, camY: number, camZ: number, yaw: number, pitch: number, fovY: number, aspect: number): void",
+            "Sets the 3D camera for this frame (fly cam: position + yaw/pitch + vertical FOV + aspect). Builds the view-projection.",
+            __RTS_FN_NS_EGUI_SET_CAMERA as *const u8,
+        ))
+        .member(func(
+            "setCameraLookAt",
+            "__RTS_FN_NS_EGUI_SET_CAMERA_LOOKAT",
+            Sig::new(vec![U64, F64, F64, F64, F64, F64, F64, F64, F64, F64, F64], AbiType::Void),
+            "setCameraLookAt(h: number, eyeX: number, eyeY: number, eyeZ: number, targetX: number, targetY: number, targetZ: number, fovY: number, aspect: number, near: number, far: number): void",
+            "Sets the 3D camera as a NaN-safe look-at (eye position looking at a target, up = +Y) with explicit near/far. More convenient than yaw/pitch for framing an object; degrades gracefully when looking straight up/down instead of gimbal-locking.",
+            __RTS_FN_NS_EGUI_SET_CAMERA_LOOKAT as *const u8,
+        ))
+        .member(func(
+            "setClearColor",
+            "__RTS_FN_NS_EGUI_SET_CLEAR_COLOR",
+            Sig::new(vec![U64, F64, F64, F64], AbiType::Void),
+            "setClearColor(h: number, r: number, g: number, b: number): void",
+            "Sets a FLAT background color (r,g,b in 0..1) for the 3D scene pass, disabling the procedural skybox. Ideal for an editor viewport that wants a neutral background instead of the starfield.",
+            __RTS_FN_NS_EGUI_SET_CLEAR_COLOR as *const u8,
+        ))
+        .member(func(
+            "setSkybox",
+            "__RTS_FN_NS_EGUI_SET_SKYBOX",
+            Sig::new(vec![U64, I64], AbiType::Void),
+            "setSkybox(h: number, on: number): void",
+            "Re-enables the procedural skybox (on!=0), undoing a previous setClearColor.",
+            __RTS_FN_NS_EGUI_SET_SKYBOX as *const u8,
+        ))
+        .member(func(
+            "setLight",
+            "__RTS_FN_NS_EGUI_SET_LIGHT",
+            Sig::new(vec![U64, F64, F64, F64, F64], AbiType::Void),
+            "setLight(h: number, dirX: number, dirY: number, dirZ: number, ambient: number): void",
+            "Sets the directional light (direction + ambient 0..1) for the 3D scene pass.",
+            __RTS_FN_NS_EGUI_SET_LIGHT as *const u8,
+        ))
+        .member(func(
+            "winWidth",
+            "__RTS_FN_NS_EGUI_WIN_WIDTH",
+            Sig::new(vec![U64], F64),
+            "winWidth(h: number): number",
+            "Current LOGICAL width (points) of the window — follows resize.",
+            __RTS_FN_NS_EGUI_WIN_WIDTH as *const u8,
+        ))
+        .member(func(
+            "winHeight",
+            "__RTS_FN_NS_EGUI_WIN_HEIGHT",
+            Sig::new(vec![U64], F64),
+            "winHeight(h: number): number",
+            "Current LOGICAL height (points) of the window — follows resize.",
+            __RTS_FN_NS_EGUI_WIN_HEIGHT as *const u8,
+        ))
+        .member(func(
+            "setShadow",
+            "__RTS_FN_NS_EGUI_SET_SHADOW",
+            Sig::new(vec![U64, F64, F64, F64, F64, F64, F64, F64], AbiType::Void),
+            "setShadow(h: number, dirX: number, dirY: number, dirZ: number, cX: number, cY: number, cZ: number, radius: number): void",
+            "Configures the directional shadow map: light travel direction (dirX/Y/Z), the center and radius of the orthographic box the shadows cover. radius<=0 disables shadows.",
+            __RTS_FN_NS_EGUI_SET_SHADOW as *const u8,
+        ))
+        .member(func(
+            "textureUpload",
+            "__RTS_FN_NS_EGUI_TEXTURE_UPLOAD",
+            Sig::new(vec![U64, U64, I64, I64], U64),
+            "textureUpload(h: number, pixelsPtr: number, w: number, h2: number): number",
+            "Uploads an RGBA8 image (pixelsPtr -> w*h*4 bytes, sRGB) to VRAM and returns a texture id (>=2) for drawMesh(..., tex=id). Typical flow: fs reads the file, imgdec.decode gives RGBA + w/h, this uploads to GPU. Sampled triplanar (world-space) in the 3D pass. 0 if invalid / non-wgpu window.",
+            __RTS_FN_NS_EGUI_TEXTURE_UPLOAD as *const u8,
+        ))
+        .member(func(
+            "drawMesh",
+            "__RTS_FN_NS_EGUI_DRAW_MESH",
+            Sig::new(vec![U64, U64, F64, F64, F64, F64, F64, F64, F64, F64, I64, I64, I64], AbiType::Void),
+            "drawMesh(h: number, meshId: number, px: number, py: number, pz: number, rx: number, ry: number, sx: number, sy: number, sz: number, color: number, emissive: number, tex: number): void",
+            "Queues a 3D draw of meshId with a transform (pos/rot euler/scale), color 0xAARRGGBB, emissive flag (0/1), and procedural texture (0=none, 1=world-space checker). Rendered in the scene pass at endFrame. Light is a POINT light at setLight's position; shadows via setShadow.",
+            __RTS_FN_NS_EGUI_DRAW_MESH as *const u8,
+        ))
+        .member(func(
+            "drawWater",
+            "__RTS_FN_NS_EGUI_DRAW_WATER",
+            Sig::new(vec![U64, U64, U64, I64, F64], I64),
+            "drawWater(h: number, meshId: number, gbuf: number, count: number, scale: number): number",
+            "Instanced water: draws `count` instances of meshId reading each instance (vec4 f32: xyz center, w signed density; w<0 = culled shell) DIRECTLY from the rts:gpu buffer `gbuf` — zero readback, one draw call. `scale` is the particle draw radius. 1 ok, 0 invalid buffer/window.",
+            __RTS_FN_NS_EGUI_DRAW_WATER as *const u8,
+        ))
+        .done();
+}
