+++ crates/rts-egui/src/scene_api.rs
@@ -29,8 +29,7 @@ fn with_scene<R: Copy>(win: u64, f: impl FnOnce(&mut Scene3D, &wgpu::Device) -> 
 
 /// Sobe uma mesh (verts interleaved pos+normal+uv = 8 f32/vértice; índices u32)
 /// pra VRAM. Retorna o id da mesh (0 se a janela não é wgpu).
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_MESH_UPLOAD(
+pub fn mesh_upload(
     win: u64,
     vptr: u64,
     vcount: i64,
@@ -46,14 +45,13 @@ pub extern "C" fn __RTS_FN_NS_EGUI_MESH_UPLOAD(
 }
 
 /// Libera uma mesh da VRAM.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_MESH_FREE(win: u64, mesh: u64) {
+pub fn mesh_free(win: u64, mesh: u64) {
     with_scene(win, |s, _d| s.free_mesh(mesh), ());
 }
 
 /// Define a câmera do frame (fly: yaw/pitch/fov/aspect). Constrói view·proj.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_CAMERA(
+#[allow(clippy::too_many_arguments)]
+pub fn set_camera(
     win: u64,
     camx: f64,
     camy: f64,
@@ -72,8 +70,8 @@ pub extern "C" fn __RTS_FN_NS_EGUI_SET_CAMERA(
 /// Câmera LOOK-AT (olho→alvo) com `near`/`far` explícitos — NaN-safe (não trava
 /// no gimbal ao olhar reto pra cima/baixo). Mais conveniente que yaw/pitch pro
 /// "frame selected" do editor. Enxertado do gpu3d (origin/main), adaptado à base LH.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_CAMERA_LOOKAT(
+#[allow(clippy::too_many_arguments)]
+pub fn set_camera_lookat(
     win: u64,
     ex: f64,
     ey: f64,
@@ -99,20 +97,17 @@ pub extern "C" fn __RTS_FN_NS_EGUI_SET_CAMERA_LOOKAT(
 
 /// Fundo CHAPADO do scene pass (r,g,b em 0..1) — desliga o skybox procedural.
 /// Ideal pro viewport do editor, que quer um fundo neutro em vez do starfield.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_CLEAR_COLOR(win: u64, r: f64, g: f64, b: f64) {
+pub fn set_clear_color(win: u64, r: f64, g: f64, b: f64) {
     with_scene(win, |s, _d| s.set_clear_color([r as f32, g as f32, b as f32, 1.0]), ());
 }
 
 /// Religa o skybox procedural (`on!=0`) desfazendo um `setClearColor`.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_SKYBOX(win: u64, on: i64) {
+pub fn set_skybox(win: u64, on: i64) {
     with_scene(win, |s, _d| s.set_skybox(on != 0), ());
 }
 
 /// Define a luz direcional (dir + ambiente).
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_LIGHT(win: u64, dx: f64, dy: f64, dz: f64, ambient: f64) {
+pub fn set_light(win: u64, dx: f64, dy: f64, dz: f64, ambient: f64) {
     with_scene(
         win,
         |s, _d| s.set_light([dx as f32, dy as f32, dz as f32], ambient as f32),
@@ -121,8 +116,7 @@ pub extern "C" fn __RTS_FN_NS_EGUI_SET_LIGHT(win: u64, dx: f64, dy: f64, dz: f64
 }
 
 /// Largura LÓGICA (points) atual da janela — segue o resize. 0 se não existe.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_WIN_WIDTH(win: u64) -> f64 {
+pub fn win_width(win: u64) -> f64 {
     crate::ctx::with_ctx(win, |c| {
         let s = c.window.inner_size();
         let sf = c.window.scale_factor();
@@ -132,8 +126,7 @@ pub extern "C" fn __RTS_FN_NS_EGUI_WIN_WIDTH(win: u64) -> f64 {
 }
 
 /// Altura LÓGICA (points) atual da janela — segue o resize. 0 se não existe.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_WIN_HEIGHT(win: u64) -> f64 {
+pub fn win_height(win: u64) -> f64 {
     crate::ctx::with_ctx(win, |c| {
         let s = c.window.inner_size();
         let sf = c.window.scale_factor();
@@ -144,8 +137,8 @@ pub extern "C" fn __RTS_FN_NS_EGUI_WIN_HEIGHT(win: u64) -> f64 {
 
 /// Configura o shadow map: direção da luz (dx,dy,dz = para onde a luz viaja) +
 /// centro (cx,cy,cz) e raio da caixa que a sombra cobre. radius<=0 desliga.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_SET_SHADOW(
+#[allow(clippy::too_many_arguments)]
+pub fn set_shadow(
     win: u64,
     dx: f64,
     dy: f64,
@@ -165,8 +158,8 @@ pub extern "C" fn __RTS_FN_NS_EGUI_SET_SHADOW(
 /// Enfileira 1 draw da mesh `mesh` com transform (pos/rot/escala) + cor
 /// (0xAARRGGBB) + emissivo (0/1) + textura procedural (0=nenhuma, 1=xadrez).
 /// O draw acontece no scene pass, no `endFrame`.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_DRAW_MESH(
+#[allow(clippy::too_many_arguments)]
+pub fn draw_mesh(
     win: u64,
     mesh: u64,
     px: f64,
@@ -201,14 +194,13 @@ pub extern "C" fn __RTS_FN_NS_EGUI_DRAW_MESH(
 /// instância (vec4 f32: xyz centro, w densidade assinada) DIRETO do buffer
 /// `gbuf` do rts:gpu — zero readback, zero FFI por partícula, 1 draw call.
 /// `scale` = raio de desenho. 1 ok, 0 = buffer/janela inválidos.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_DRAW_WATER(
-    win: u64,
-    mesh: u64,
-    gbuf: u64,
-    count: i64,
-    scale: f64,
-) -> i64 {
+///
+/// Só existe com `old-engine`: `gbuf` é um handle do `rts:gpu`, cujos buffers
+/// vivem no `HandleTable` do motor antigo. Sem ele não há de onde ler as
+/// instâncias — e devolver 0 caladamente seria uma água que não aparece sem
+/// dizer por quê.
+#[cfg(feature = "old-engine")]
+pub fn draw_water(win: u64, mesh: u64, gbuf: u64, count: i64, scale: f64) -> i64 {
     if count <= 0 {
         return 0;
     }
@@ -228,8 +220,7 @@ pub extern "C" fn __RTS_FN_NS_EGUI_DRAW_WATER(
 /// Sobe uma imagem RGBA8 (`ptr` → w*h*4 bytes, sRGB) pra VRAM e devolve um id de
 /// textura (>=2) usável em `drawMesh(..., tex=id)`. Fluxo típico: `fs` lê o arquivo,
 /// `imgdec.decode` devolve RGBA + w/h, e isto sobe pra GPU. 0 se inválido/não-wgpu.
-#[unsafe(no_mangle)]
-pub extern "C" fn __RTS_FN_NS_EGUI_TEXTURE_UPLOAD(win: u64, ptr: u64, w: i64, h: i64) -> u64 {
+pub fn texture_upload(win: u64, ptr: u64, w: i64, h: i64) -> u64 {
     if ptr == 0 || w <= 0 || h <= 0 {
         return 0;
     }
