+++ crates/rts-render/src/lib.rs
@@ -14,7 +14,9 @@
 
 use std::cell::RefCell;
 
+#[cfg(feature = "old-engine")]
 pub mod abi;
+#[cfg(feature = "old-engine")]
 pub use abi::register;
 
 /// Prelude `.ts` da fachada ERGONÔMICA `rts:canvas` — UI imediata (Canvas com
