+++ crates/rts-runtime/Cargo.toml
@@ -58,6 +58,10 @@ rts-input = { path = "../rts-input" }
 # abaixo: dx12=Windows, vulkan=Linux, metal=macOS. (Linux também precisa do winit
 # x11/wayland — resolvido no Cargo.toml do rts-egui.) Features unem por target.
 rts-egui = { path = "../rts-egui", default-features = false, features = [
+    # `old-engine`: os símbolos `__RTS_FN_NS_EGUI_*` e o namespace do motor
+    # ANTIGO. Esta facade É o motor antigo, então é ela quem os pede — o crate
+    # rts-egui não os assume mais. O motor novo consome o mesmo crate sem isto.
+    "old-engine",
     "wgpu-backend",
     "glow-backend",
 ] }
